polish parallel6 example
